Add change user status

+++ controllers/feed.js
@@ -116,6 +116,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator !== req.userId) {
+        const error = new Error('Not Authorized!')
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl)
       }
@@ -144,12 +149,23 @@ exports.deletePost = (req, res, next) => {
           error.statusCode = 404;
           throw error;
         }
+        if (post.creator !== req.userId) {
+          const error = new Error('Not Authorized!')
+          error.statusCode = 403;
+          throw error;
+        }
         clearImage(post.imageUrl);
         return Post.findByIdAndRemove(post._id);
       })
       .then(result => {
-        console.log(result)
-        res.status(200).json({message: 'Post deleted'})
+        return User.findById(req.userId)
+      })
+      .then(user => {
+        user.posts.pull(postId);
+        return user.save();
+      })
+      .then(result => {
+        res.status(200).json({message: 'Post deleted'});
       })
       .catch(err => {
         if (!err.statusCode) {
